+++ blockchain/v2/types.go
@@ -1,35 +1,4 @@
 package v2
 
-import "time"
-
 type Event interface{}
-
 type Events []Event
-type testEvent struct {
-	msg  string
-	time time.Time
-}
-
-type testEventTwo struct {
-	msg string
-}
-
-type stopEvent struct{}
-type timeCheck struct {
-	time time.Time
-}
-
-type errEvent struct{}
-
-type scTestEvent struct{}
-
-type pcFinished struct{}
-
-type routineFinished struct{}
-
-func (rf *routineFinished) Error() string {
-	return "routine finished"
-}
-
-type scFull struct{}
-type pcFull struct{}
